requirements.txt
openai
python-dotenv
gradio
PyPDF2
streamlit
azure-kusto-data
pyautogen